Refactored cron logic
