+++ .github/workflows/lint.yaml
@@ -14,7 +14,7 @@ jobs:
         uses: actions/checkout@v5
 
       - name: Setup Go
-        uses: actions/setup-go@v6
+        uses: actions/Setup-go@v6
         with:
           go-version-file: 'go.mod'
           cache: true
