--- internal/cron/model/model.go
@@ -1,9 +1,0 @@
-package model
-
-import "context"
-
-type CronLike interface {
-	Add(ctx context.Context) error
-	Start()
-	Stop() context.Context
-}
